move cache classes
