+++ src/main/java/bdv/cache/CacheHints.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,12 +26,13 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package bdv.img.cache;
+package bdv.cache;
 
+import bdv.cache.util.BlockingFetchQueues;
 import bdv.img.cache.VolatileGlobalCellCache.VolatileCellCache;
 
 /**
- * Hints to the {@link VolatileCellCache cache} on how to handle cell requests.
+ * Hints to the {@link LoadingVolatileCache cache} on how to handle cell requests.
  * Consists of the {@link LoadingStrategy} for cells, the priority with which to
  * enqueue cells to the {@link BlockingFetchQueues} (if they are enqueued) and
  * whether they should be enqueued to the front (most recent requests are
