+++ src/main/java/bdv/img/cache/CachedCellImg.java
@@ -32,6 +32,7 @@ import net.imglib2.img.basictypeaccess.volatiles.VolatileAccess;
 import net.imglib2.img.cell.AbstractCellImg;
 import net.imglib2.img.cell.CellImgFactory;
 import net.imglib2.type.NativeType;
+import bdv.cache.CacheHints;
 import bdv.img.cache.VolatileImgCells.CellCache;
 
 final public class CachedCellImg< T extends NativeType< T >, A extends VolatileAccess >
