+++ src/main/java/bdv/img/openconnectome/OpenConnectomeImageLoader.java
@@ -39,10 +39,10 @@ import com.google.gson.JsonSyntaxException;
 import bdv.AbstractViewerSetupImgLoader;
 import bdv.ViewerImgLoader;
 import bdv.ViewerSetupImgLoader;
-import bdv.img.cache.Cache;
-import bdv.img.cache.CacheHints;
+import bdv.cache.Cache;
+import bdv.cache.CacheHints;
+import bdv.cache.LoadingStrategy;
 import bdv.img.cache.CachedCellImg;
-import bdv.img.cache.LoadingStrategy;
 import bdv.img.cache.VolatileGlobalCellCache;
 import bdv.img.cache.VolatileImgCells;
 import bdv.img.cache.VolatileImgCells.CellCache;
