+++ src/main/java/bdv/img/remote/RemoteImageLoader.java
@@ -47,9 +47,9 @@ import net.imglib2.util.IntervalIndexer;
 import net.imglib2.view.Views;
 import bdv.AbstractViewerSetupImgLoader;
 import bdv.ViewerImgLoader;
-import bdv.img.cache.CacheHints;
+import bdv.cache.CacheHints;
+import bdv.cache.LoadingStrategy;
 import bdv.img.cache.CachedCellImg;
-import bdv.img.cache.LoadingStrategy;
 import bdv.img.cache.VolatileGlobalCellCache;
 import bdv.img.cache.VolatileImgCells;
 import bdv.img.cache.VolatileImgCells.CellCache;
