+++ src/main/java/bdv/spimdata/legacy/LegacyViewerImgLoader.java
@@ -28,12 +28,12 @@
  */
 package bdv.spimdata.legacy;
 
+import bdv.cache.Cache;
 import mpicbg.spim.data.legacy.LegacyBasicImgLoader;
 import mpicbg.spim.data.sequence.ViewId;
 import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.Volatile;
 import net.imglib2.realtransform.AffineTransform3D;
-import bdv.img.cache.Cache;
 
 //@Deprecated
 public interface LegacyViewerImgLoader< T, V extends Volatile< T > > extends LegacyBasicImgLoader< T >
