+++ src/main/java/bdv/spimdata/legacy/LegacyViewerImgLoaderWrapper.java
@@ -37,7 +37,7 @@ import net.imglib2.Volatile;
 import net.imglib2.realtransform.AffineTransform3D;
 import bdv.ViewerImgLoader;
 import bdv.ViewerSetupImgLoader;
-import bdv.img.cache.Cache;
+import bdv.cache.Cache;
 
 //@Deprecated
 public class LegacyViewerImgLoaderWrapper< T, V extends Volatile< T >, I extends LegacyViewerImgLoader< T, V > > implements ViewerImgLoader
