+++ src/main/java/bdv/tools/RecordMaxProjectionDialog.java
@@ -70,8 +70,8 @@ import net.imglib2.ui.OverlayRenderer;
 import net.imglib2.ui.PainterThread;
 import net.imglib2.ui.RenderTarget;
 import net.imglib2.util.LinAlgHelpers;
+import bdv.cache.Cache;
 import bdv.export.ProgressWriter;
-import bdv.img.cache.Cache;
 import bdv.util.Prefs;
 import bdv.viewer.ViewerPanel;
 import bdv.viewer.overlay.ScaleBarOverlayRenderer;
