+++ src/main/java/bdv/tools/RecordMovieDialog.java
@@ -63,8 +63,8 @@ import net.imglib2.realtransform.AffineTransform3D;
 import net.imglib2.ui.OverlayRenderer;
 import net.imglib2.ui.PainterThread;
 import net.imglib2.ui.RenderTarget;
+import bdv.cache.Cache;
 import bdv.export.ProgressWriter;
-import bdv.img.cache.Cache;
 import bdv.util.Prefs;
 import bdv.viewer.ViewerPanel;
 import bdv.viewer.overlay.ScaleBarOverlayRenderer;
