+++ src/main/java/bdv/viewer/render/MipmapOrdering.java
@@ -31,8 +31,8 @@ package bdv.viewer.render;
 import java.util.Comparator;
 import java.util.List;
 
+import bdv.cache.CacheHints;
 import net.imglib2.realtransform.AffineTransform3D;
-import bdv.img.cache.CacheHints;
 
 public interface MipmapOrdering
 {
